Add password recovery process
Email the user with a link, with a token, and save it in a new table, password_recovery. The token has a 24h and can only be used 1 time.
passwordRecovery end point add to UserController, used passwordRecovery in user service
passwordRecovery check the id in the url, the token and the expiration date, later take the new password and validate it and updated into the user table.
All the password recovery end points were habilitate to be used by all user, without authorization